refactor: LottoWheel -> LottoNumberCache로 변경
- LottoNumberCache: 캐싱해둔 LottoNumber를 꺼내주는 역할
- 로또 생성은 Lotto 클래스 내부에서 처리
