+++ src/main/java/lotto/model/Lotto.java
@@ -4,6 +4,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import lotto.model.number.LottoNumber;
+import lotto.model.number.LottoNumberCache;
 
 /*
  * 로또 한 장을 의미하는 Class
@@ -44,6 +45,17 @@ public class Lotto {
         return (int) numbers.stream()
                 .distinct()
                 .count();
+    }
+
+    public static Lotto ofRandom() {
+        return new Lotto(LottoNumberCache.getRandomNumbers(NUMBERS_SIZE), true);
+    }
+
+    public static Lotto from(List<String> inputs) {
+        List<LottoNumber> numbers = inputs.stream()
+                .map(LottoNumberCache::getNumber)
+                .collect(Collectors.toList());
+        return new Lotto(numbers, false);
     }
 
     public boolean contains(LottoNumber number) {
