+++ src/test/java/lotto/model/number/LottoNumberCacheTest.java
@@ -1,0 +1,34 @@
+package lotto.model.number;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+
+public class LottoNumberCacheTest {
+    @DisplayName("로또 번호가 숫자가 아니면 예외가 발생한다")
+    @Test
+    void type_exception() {
+        assertThatThrownBy(() -> LottoNumberCache.getNumber("일"))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("[ERROR] 로또 번호는 숫자로만 입력해주세요");
+    }
+
+    @ParameterizedTest(name = "{0}에 해당하는 LottoNumber를 반환한다")
+    @ValueSource(ints = {1, 15, 40, 23})
+    void get_number(int number) {
+        assertThat(LottoNumberCache.getNumber(String.valueOf(number)).getValue()).isEqualTo(number);
+    }
+
+    @DisplayName("랜덤 로또를 추출한 후에도 알맞는 LottoNumber를 반환한다")
+    @ParameterizedTest(name = "{0}에 해당하는 LottoNumber를 반환한다")
+    @ValueSource(ints = {1, 15, 40, 23})
+    void get_number_after_shuffle(int number) {
+        LottoNumberCache.getRandomNumbers(1);
+
+        assertThat(LottoNumberCache.getNumber(String.valueOf(number)).getValue()).isEqualTo(number);
+    }
+}
